Code cleanup; change filenames
